requirements.txt
flask
azure-identity
azure-storage-blob
gunicorn
PyJWT